+++ src/App.tsx
@@ -1,7 +1,7 @@
 import { BadgeText } from '@/components/ui/Text'
-import Card from '@/components/ui/Card'
 import FooterBlog from '@/components/Footer/FooterBlog'
 import Navbar from '@/components/navbar/Navbar'
+import PostCard from '@/components/ui/PostCard'
 
 export default function () {
   return (
@@ -13,11 +13,11 @@ export default function () {
         logoText="SealCaster"
         noWalletText={'Connect burner wallet'}
       />
-      <Card>
+      <PostCard>
         <BadgeText>
           <FooterBlog />
         </BadgeText>
-      </Card>
+      </PostCard>
     </div>
   )
 }
